+++ src/vue/modules/create-sale-form/components/short-blurb-step-form.vue
@@ -50,7 +50,8 @@
 import FormMixin from '@/vue/mixins/form.mixin'
 
 import { DOCUMENT_TYPES } from '@/js/const/document-types.const'
-import { DocumentContainer } from '@/js/helpers/DocumentContainer'
+import { DOCUMENT_POLICIES } from '@/js/const/document-policies.const'
+import { Document } from '@tokend/js-sdk'
 
 import { CreateSaleRequest } from '../wrappers/create-sale-request'
 
@@ -101,7 +102,10 @@ export default {
     populateForm () {
       this.form = {
         saleLogo: this.request.logoKey
-          ? new DocumentContainer(this.request.logo)
+          ? new Document(
+            this.request.logo,
+            DOCUMENT_POLICIES[this.request.logo.type],
+          )
           : null,
         shortDescription: this.request.shortDescription,
       }