+++ src/vue/modules/verification/general-form/store/index.js
@@ -1,13 +1,12 @@
 import { DOCUMENT_TYPES } from '@/js/const/document-types.const'
 import { ID_DOCUMENT_TYPES } from '../id-document-types'
-import { BLOB_TYPES } from '@tokend/js-sdk'
+import { BLOB_TYPES, Document } from '@tokend/js-sdk'
 import { types } from './types'
 import { api } from '@/api'
 
 import { wrapDocument } from './wrap-document'
 
 import { toRFC3339, fromRFC3339 } from '../format-date'
-import { uploadDocument } from '@/js/helpers/upload-documents'
 
 const state = {
   isAccredited: false,
@@ -149,13 +148,12 @@ const actions = {
   },
 
   async [types.UPLOAD_DOCUMENTS] ({ state }) {
-    await Promise.all([
-      // it modifies the state intentionally
-      uploadDocument(state.documents.avatar),
-      uploadDocument(state.documents.selfie),
-      uploadDocument(state.documents.idDocumentFace),
-      uploadDocument(state.documents.idDocumentBack),
-      uploadDocument(state.documents.proofOfInvestor),
+    await Document.uploadDocuments([
+      state.documents.avatar,
+      state.documents.selfie,
+      state.documents.idDocumentFace,
+      state.documents.idDocumentBack,
+      state.documents.proofOfInvestor,
     ])
   },
 
@@ -229,36 +227,36 @@ const getters = {
     if (state.isAccredited) {
       if (state.kycProofOfInvestor) {
         blobData.documents[DOCUMENT_TYPES.kycProofOfInvestor] =
-          state.documents.proofOfInvestor.getDetailsForSave()
+          state.documents.proofOfInvestor.toJSON()
       }
     }
 
     if (state.documents.idDocumentFace) {
       blobData.documents[DOCUMENT_TYPES.kycIdDocument].face =
-        state.documents.idDocumentFace.getDetailsForSave()
+        state.documents.idDocumentFace.toJSON()
     }
 
     if (state.documents.selfie) {
       blobData.documents[DOCUMENT_TYPES.kycSelfie] =
-        state.documents.selfie.getDetailsForSave()
+        state.documents.selfie.toJSON()
     }
 
     // avatar is not required, or it may not exist in old kyc data
     if (state.documents.avatar) {
       blobData.documents[DOCUMENT_TYPES.kycAvatar] =
-        state.documents.avatar.getDetailsForSave()
+        state.documents.avatar.toJSON()
     }
 
     // back side of documents is needed only for specific document types
     if (state.documents.idDocumentBack) {
       blobData.documents[DOCUMENT_TYPES.kycIdDocument].back =
-        state.documents.idDocumentBack.getDetailsForSave()
+        state.documents.idDocumentBack.toJSON()
     }
 
     // is required only when applying for accredited investor role
     if (state.documents.proofOfInvestor) { // TODO
       blobData.documents[DOCUMENT_TYPES.kycProofOfInvestor] =
-        state.documents.proofOfInvestor.getDetailsForSave()
+        state.documents.proofOfInvestor.toJSON()
     }
 
     return blobData